+++ tabletest-reporter-gradle-plugin/src/test/java/io/github/nchaugen/tabletest/gradle/TableTestReporterPluginTest.java
@@ -52,13 +52,13 @@ class TableTestReporterPluginTest {
         Path tableDir = testClassDir.resolve("leapYearRules(java.time.Year, boolean)");
         Files.createDirectories(tableDir);
 
-        Files.writeString(testClassDir.resolve("Calendar Calculations.yaml"), """
+        Files.writeString(testClassDir.resolve("TABLETEST-Calendar Calculations.yaml"), """
             "title": "Calendar"
             "description": "Various rules for calendar calculations."
             """
         );
 
-        Files.writeString(tableDir.resolve("Leap Year Rules.yaml"), """
+        Files.writeString(tableDir.resolve("TABLETEST-Leap Year Rules.yaml"), """
             "title": "Leap Year Rules with Single Example"
             "description": "The leap year rules should be well-known."
             "headers":